requirements.txt
django
gunicorn
whitenoise
psycopg2-binary
requests
urllib3
django-environ
djangorestframework
django-cors-headers
spotipy